[Update] Changing Column Types in Migration Files

+++ db/schema.rb
@@ -63,7 +63,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_11_16_005041) do
   end
 
   create_table "sessions", force: :cascade do |t|
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.string "ip_address"
     t.string "user_agent"
     t.datetime "created_at", null: false
